+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @duck4i/retro-ui
 
+## 0.3.3
+
+### Patch Changes
+
+- input fixes and inputbox adde
+
 ## 0.3.2
 
 ### Patch Changes